feat(#47): simple commenting has been implemented

+++ ProjectService/CommentAPI/Controllers/CommentController.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using MongoDB.Driver;
 using System;
 using System.Collections.Generic;
+using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -23,13 +24,23 @@ namespace CommentAPI.Controllers
         }
 
         [HttpPost]
-        public async Task<ActionResult<Comment>> Create(CommentCreateDto commentToCreate)
+        public async Task<ActionResult<Comment>> Create(CommentCreateDto commentToCreate, [FromHeader] string authorization)
         {
+            if (authorization == null)
+                return BadRequest("Missing authorization header!");
+
+            string token = ParseToken(authorization);
+            if (token == null)
+                return BadRequest("Wrong type of authorization header!");
+
+            var jwtToken = new JwtSecurityToken(token);
+
             var comment = new Comment
             {
                 ToDoId = commentToCreate.ToDoId,
                 Content = commentToCreate.Content,
-                User = "Unknown" // TODO get user id from token
+                User = jwtToken.Subject,
+                CreationTime = DateTime.Now
             };
 
             try
@@ -65,14 +76,26 @@ namespace CommentAPI.Controllers
         }
 
         [HttpPut]
-        public async Task<IActionResult> UpdateContent(CommentUpdateDto commentToUpdate)
+        public async Task<IActionResult> UpdateContent(CommentUpdateDto commentToUpdate, [FromHeader] string authorization)
         {
+            if (authorization == null)
+                return BadRequest("Missing authorization header!");
+
+            string token = ParseToken(authorization);
+            if (token == null)
+                return BadRequest("Wrong type of authorization header!");
+
+            var jwtToken = new JwtSecurityToken(token);
+
             var comment = await _repository.Get(commentToUpdate.CommentId);
 
             if (comment == null)
             {
                 return NotFound("Comment with " + commentToUpdate.CommentId + " not found");
             }
+
+            if (jwtToken.Subject != comment.User)
+                return Unauthorized("Can't modify other user's comments");
 
             comment.Content = commentToUpdate.newContent;
             await _repository.Update(comment);
@@ -81,8 +104,17 @@ namespace CommentAPI.Controllers
         }
 
         [HttpDelete("{id:length(24)}")]
-        public async Task<IActionResult> Delete(string id)
+        public async Task<IActionResult> Delete(string id, [FromHeader] string authorization)
         {
+            if (authorization == null)
+                return BadRequest("Missing authorization header!");
+
+            string token = ParseToken(authorization);
+            if (token == null)
+                return BadRequest("Wrong type of authorization header!");
+
+            var jwtToken = new JwtSecurityToken(token);
+
             var comment = await _repository.Get(id);
 
             if (comment == null)
@@ -90,10 +122,21 @@ namespace CommentAPI.Controllers
                 return NotFound("Comment with " + id + " not found");
             }
 
+            if (jwtToken.Subject != comment.User)
+                return Unauthorized("Can't modify other user's comments");
+
             await _repository.Delete(id);
 
             return Ok();
         }
 
+        private string ParseToken(string authHeader)
+        {
+            string[] parts = authHeader.Split(null);
+            if (parts.Length != 2 || parts[0] != "Bearer")
+                return null;
+            else
+                return parts[1];
+        }
     }
 }
